Make input file & manual fields optional
Reworks the batch_create to allow for modules that don't need an input file and adds --manual option that allows user to specify manual field creation vs using fields in the modules dict.
Manual is mainly a testing feature for unsupported modules and the modules dict will be the official way to specify fields.